pyproject.toml
[project]
name = "megapolis-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Rishabh",email = "rishabhgautam727@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "hypercorn (>=0.16.0,<0.17.0)",
    "SQLAlchemy (>=2.0.32,<3.0.0)",
    "psycopg[binary,pool] (>=3.2.1,<4.0.0)",
    "alembic (>=1.13.2,<2.0.0)",
    "typer (>=0.12.3,<0.13.0)",
    "click (>=8.1.3,<8.2.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "supabase (>=2.18.1,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "loguru (>=0.7.2,<0.8.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "aiosmtplib (>=4.0.2,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-genai (>=1.32.0,<2.0.0)",
    "infisicalsdk (>=1.0.11,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
anyio = "^4.10.0"
httpx = "^0.28.1"
aiosqlite = "^0.21.0"

